renamed sample_component.py to sample.py
