+++ sample.py
